Refactor entrypoint for better UX
